requirements.txt
pandas==2.2.2
plyer==2.1.0
pyinstaller==6.10.0
openpyxl==3.1.5
requests==2.32.3
pip-audit==2.7.3
python-telegram-bot==21.5
